requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg==3.1.12
redis==5.0.1
celery==5.3.4
django-cors-headers==4.3.0
